refactor: group all Login dependecies in a single file
