+++ src/presentation/controllers/login/login-protocols.ts
@@ -1,0 +1,3 @@
+export * from '../../protocols'
+export * from '../../protocols/email-validator'
+export * from '../../../domain/usecases/authentication'
